Move inline components into new folder strucutre. Create placeholder tests.
