+++ src/inlines/Inline/Inline.test.tsx
@@ -1,0 +1,7 @@
+/* eslint func-names: 0 */
+
+import expect from 'expect'
+
+describe('Inline: to do', function () {
+  it('has no tests yet')
+})
