+++ src/inlines/Text/Text.test.tsx
@@ -1,0 +1,7 @@
+/* eslint func-names: 0 */
+
+import expect from 'expect'
+
+describe('Text: to do', function () {
+  it('has no tests yet')
+})
